add delete specs, owner? method, remove expose

+++ app/controllers/answers_controller.rb
@@ -1,34 +1,53 @@
 class AnswersController < ApplicationController
   before_action :authenticate_user!, except: %i[index show]
-  expose(:question)
-  expose(:answers, ancestor: :question)
-  expose(:answer)
+
+  def show; end
+
+  def new; end
+
+  def edit; end
 
   def create
     @answer = question.answers.new(answer_params)
+    @answer.author_id = current_user.id
     if @answer.save
       redirect_to @answer.question, notice: 'Your answer has been successfully added.'
     else
-      redirect_to @answer.question, alert: 'Your answer has not been saved!'
+      redirect_to @answer.question, alert: 'Your answer has not been saved! ' + @answer.errors.full_messages.first.to_s
     end
   end
 
   def update
     if answer.update(answer_params)
-      redirect_to answer.question
+      redirect_to @answer.question, notice: 'Your answer successful updated!'
     else
-      render :edit
+      render :edit, alert: 'Your answer has not been saved!'
     end
   end
 
   def destroy
-    answer.destroy
-    redirect_to answer.question
+    if answer.owner?(current_user)
+      answer.destroy
+      redirect_to answer.question, notice: 'Your answer successfully deleted.'
+    else
+      redirect_to answer.question, alert: "You can't delete not your question!"
+    end
   end
 
   private
 
+  def answer
+    @answer ||= params[:id] ? Answer.find(params[:id]) : Answer.new
+  end
+
+  def question
+    @question ||= Question.find(params[:question_id])
+  end
+
+  helper_method :question
+  helper_method :answer
+
   def answer_params
-    params.require(:answer).permit(:body)
+    params.require(:answer).permit(:body, :author_id)
   end
 end
